Changes according to review comments
